Numerous Updates:  MedDream, OHIF, update Osimis Image, Python scripts, Postgres support for MWL.
Postgres is not completely supported yet for the MWL files, but a work in progress.  You'll need to initialize the orthanc_ris DB in Postgres.  There are some instructions in the postgres_init.  Will try to automate that all later and clean up the config and orthanc.json to make it possible to choose none, one or both.

Downloading a .zip packages up the Radiant Viewer and ReadMe, etc.

Meddream and OHIF are integrated into Explorer2.